+++ app/routes/components/Switch/demos/WithDifferentSizes/index.tsx
@@ -2,17 +2,21 @@ import { useState } from "react";
 
 import { Flex } from "@/components/Flex";
 import { Switch } from "@/components/Switch";
+import { Typography } from "@/components/Typography";
 
 export default function Demo() {
   const [checked, setChecked] = useState(false);
 
   return (
-    <Flex dir="column" alignItems="center" gap="2rem">
-      <Flex alignItems="center" gap="0.25rem">
-        <span>Use inline Switch with text →</span>
+    <Flex dir="column" items="center" gap="2rem">
+      <Flex items="center" gap="0.25rem">
+        <span>
+          Use inline Switch with <Typography.Code>size="text"</Typography.Code>{" "}
+          →
+        </span>
         <Switch size="text" value={checked} onChange={setChecked} />
       </Flex>
-      <Flex alignItems="center" gap="1rem">
+      <Flex items="center" gap="1rem">
         <Switch size="xs" value={checked} onChange={setChecked} />
         <Switch size="sm" value={checked} onChange={setChecked} />
         <Switch size="md" value={checked} onChange={setChecked} />