+++ src/components/Switch/CHANGELOG.md
@@ -8,3 +8,4 @@
 - 2025-10-14: Add border styles, use ratio-based width calculation, adjust thumb
   gap values. 
 - 2025-11-03: Use JSS & resolve-style for styles.
+- 2025-12-07: Refactor code and remove deprecated dependencies.